chore: split models into a subpackage and per-symbol docs nav (refs #67)
Two A+ backlog items from the #67 grading sweep, both with zero public-API
change (compatibility gate passes, public-API snapshot unchanged).

models.py (864 LOC) -> models/ subpackage
  Mirrors the protocols/ per-resource split. Grouped by resource into
  private submodules: _protocols (Protocol/Capability literals, tables,
  normalize_*), _helpers (payload coercion), _discovery (ServiceSummary,
  DataPlaneCapabilities), _sources (SourceLocator, SourceDescriptor),
  _query (Pagination, Query, DegradedReason, Result), _features (Feature,
  FeatureSet, FeatureQuery, QueryFeature, FeatureQueryResult), and _edits
  (EditOperationResult, ApplyEditsResult). The models/__init__ re-exports
  every public name unchanged and pins each symbol's __module__ back to
  honua_sdk.models, so `from honua_sdk.models import X`, the top-level
  `from honua_sdk import X`, repr()/pickle identity, and the public-API
  snapshot all stay byte-for-byte stable. Wheel contents verified to
  include every new submodule.

Per-symbol API-reference nav
  Split the two monolithic mkdocstrings dumps (reference/honua_sdk.md,
  honua_admin.md) into per-section pages under reference/honua-sdk/ and
  reference/honua-admin/ (Clients / Source facade / Models / Errors), each
  with its own nav entry so the reference scales as the surface grows.
  All prose, examples, and cross-links preserved; index.md links updated.
  `mkdocs build --strict` is clean.

Verification (local): ruff clean; mypy --strict Success (61 files);
pytest 1115 passed, 1 skipped, 94.60% coverage (gate 94%); both wheels
build + twine PASSED; compatibility gate passed; mkdocs build --strict clean.

Deferred (remain open on #67): client.py/async_client.py and honua-admin
sync/async dedup into resource mixins; SourceClientProtocol factory Any
returns; auto Idempotency-Key on all user POSTs; protocols/{geoservices,
odata}.py splits; docs site/ artifact upload; CHANGELOG before/after
blocks. Held back to keep this increment risk-free and API-stable.
