+++ docs/index.md
@@ -53,7 +53,7 @@ pip install honua-admin
 | Query features in 5 minutes     | [Quickstart](quickstart.md)                           |
 | Stream features over gRPC       | [Core client](core-client.md)                         |
 | Wire a FastAPI service          | [Examples](examples.md)                               |
-| Manage services & connections   | [honua-admin reference](reference/honua_admin.md)     |
+| Manage services & connections   | [honua-admin reference](reference/honua-admin/clients.md) |
 | Understand the protocol matrix  | [Protocol parity](protocol-parity.md)                 |
 | Diagnose an error               | [Troubleshooting](troubleshooting.md)                 |
 
@@ -62,7 +62,7 @@ pip install honua-admin
 - [Quickstart](quickstart.md) -- 5-minute setup and first query.
 - [Core client guide](core-client.md) -- the `Source` / `Query` / `Result`
   facade and capability checks.
-- [API Reference](reference/honua_sdk.md) -- autogenerated docs for every
+- [API Reference](reference/honua-sdk/clients.md) -- autogenerated docs for every
   public class in `honua-sdk` and `honua-admin`.
 - [Examples](examples.md) -- runnable demos (ETL, FastAPI, data
   quality reporting).
