+++ docs/reference/honua-admin/models.md
@@ -1,32 +1,4 @@
-# honua-admin Reference
-
-## Clients
-
-The admin SDK ships sync and async clients with parity to the data-plane
-SDK: pick [`HonuaAdminClient`][honua_admin.HonuaAdminClient] for scripts
-and CLIs, and
-[`AsyncHonuaAdminClient`][honua_admin.AsyncHonuaAdminClient] for async
-services. Both clients share the data-plane configuration model
-(auth, retries, timeouts, ``with_options(...)``); see the
-[honua-sdk core client model](../core-client.md) for the conceptual
-overview that applies to both packages.
-
-```python
-from honua_admin import HonuaAdminClient
-
-with HonuaAdminClient("https://admin.your-honua-server.com", api_key="...") as admin:
-    services = admin.list_services()
-```
-
-**See also**: [honua-sdk Clients](honua_sdk.md#clients) for the data-plane
-counterpart that shares the same configuration model, and
-[Core client model](../core-client.md) for cross-package `with_options(...)`
-semantics.
-
-::: honua_admin.HonuaAdminClient
-::: honua_admin.AsyncHonuaAdminClient
-
-## Models
+# honua-admin › Models
 
 The model surface is grouped by admin domain — service catalog and
 capabilities, the metadata manifest apply/diff flow, published-layer
@@ -34,7 +6,7 @@ and style payloads, and secure-connection management. Each group mirrors
 a coherent slice of the admin REST surface, so you typically only touch
 one group per workflow.
 
-### Services and capabilities
+## Services and capabilities
 
 These types describe what an admin tenant exposes — the catalog of
 services, the resolved capability set, the running server version, and
@@ -46,7 +18,7 @@ endpoints.
 ::: honua_admin.AdminVersionResponse
 ::: honua_admin.AdminCompatibilityFeatureFlags
 
-### Metadata manifests
+## Metadata manifests
 
 The manifest types power the declarative apply/diff workflow:
 [`MetadataManifest`][honua_admin.MetadataManifest] is the input bundle,
@@ -67,7 +39,7 @@ canonical resource shape consumed and emitted by the apply API.
 ::: honua_admin.MetadataResourceIdentifier
 ::: honua_admin.ResourceMetadata
 
-### Published layers and styles
+## Published layers and styles
 
 These types describe layers exposed by the tenant and the styles
 attached to them — useful when reconciling rendering metadata or
@@ -76,7 +48,7 @@ auditing the published surface.
 ::: honua_admin.PublishedLayerSummary
 ::: honua_admin.LayerStyleResponse
 
-### Secure connections
+## Secure connections
 
 The secure-connection types cover the lifecycle of upstream credentials:
 listing and inspecting connections, testing connectivity, validating
@@ -87,21 +59,3 @@ encryption at rest, and rotating tenant keys.
 ::: honua_admin.ConnectionTestResult
 ::: honua_admin.EncryptionValidationResult
 ::: honua_admin.KeyRotationResult
-
-## Errors
-
-The admin client raises the same exception hierarchy as the data-plane
-SDK — there is no admin-specific exception surface. Catch
-``HonuaError`` for any failure, or one of its subclasses
-(``HonuaHttpError``/``HonuaAuthError``/``HonuaRateLimitError`` for HTTP
-failures; ``HonuaTransportError``/``HonuaTimeoutError`` for transport
-failures) when you need finer-grained handling. See the
-[honua-sdk error reference](honua_sdk.md#errors) for the full
-documentation of:
-
-- [`HonuaError`](honua_sdk.md#honua_sdk.errors.HonuaError)
-- [`HonuaHttpError`](honua_sdk.md#honua_sdk.errors.HonuaHttpError)
-- [`HonuaAuthError`](honua_sdk.md#honua_sdk.errors.HonuaAuthError)
-- [`HonuaRateLimitError`](honua_sdk.md#honua_sdk.errors.HonuaRateLimitError)
-- [`HonuaTransportError`](honua_sdk.md#honua_sdk.errors.HonuaTransportError)
-- [`HonuaTimeoutError`](honua_sdk.md#honua_sdk.errors.HonuaTimeoutError)
