--- packages/honua-sdk/honua_sdk/models.py
@@ -1,864 +1,0 @@
-"""Typed models for core Honua SDK responses."""
-
-from __future__ import annotations
-
-import contextlib
-from collections.abc import Iterable, Mapping, Sequence
-from dataclasses import dataclass, field
-from typing import Any, Generic, Literal, TypeAlias, TypeVar
-
-T = TypeVar("T")
-
-#: Canonical cross-SDK protocol identifier. Pass aliases through
-#: :func:`normalize_protocol` to coerce them before storing in typed code paths.
-Protocol: TypeAlias = Literal[
-    "grpc",
-    "geoservices-feature-service",
-    "geoservices-map-service",
-    "geoservices-image-service",
-    "geoservices-geometry-service",
-    "geoservices-gp-service",
-    "ogc-features",
-    "ogc-tiles",
-    "ogc-maps",
-    "stac",
-    "wfs",
-    "wms",
-    "wmts",
-    "odata",
-    "maplibre-vector",
-    "maplibre-raster",
-    "maplibre-geojson",
-]
-#: Canonical cross-SDK capability identifier. Use :func:`normalize_capability`
-#: to coerce snake_case or alias forms before passing into typed APIs.
-Capability: TypeAlias = Literal[
-    "query",
-    "queryAggregate",
-    "queryExtent",
-    "queryObjectIds",
-    "queryRelated",
-    "applyEdits",
-    "attachments",
-    "render",
-    "tiles",
-    "sql",
-    "stream",
-    "pbf",
-    "connect",
-    "image",
-    "geometry",
-    "geoprocess",
-    "processes",
-]
-QueryProtocol: TypeAlias = Protocol
-
-PROTOCOLS = (
-    "grpc",
-    "geoservices-feature-service",
-    "geoservices-map-service",
-    "geoservices-image-service",
-    "geoservices-geometry-service",
-    "geoservices-gp-service",
-    "ogc-features",
-    "ogc-tiles",
-    "ogc-maps",
-    "stac",
-    "wfs",
-    "wms",
-    "wmts",
-    "odata",
-    "maplibre-vector",
-    "maplibre-raster",
-    "maplibre-geojson",
-)
-
-PROTOCOL_ALIASES: Mapping[str, str] = {
-    "feature-server": "geoservices-feature-service",
-    "featureserver": "geoservices-feature-service",
-    "feature-service": "geoservices-feature-service",
-    "geoservices-featureserver": "geoservices-feature-service",
-    "map-server": "geoservices-map-service",
-    "mapserver": "geoservices-map-service",
-    "image-server": "geoservices-image-service",
-    "imageserver": "geoservices-image-service",
-    "geometry-server": "geoservices-geometry-service",
-    "geometryserver": "geoservices-geometry-service",
-    "gp-server": "geoservices-gp-service",
-    "gpserver": "geoservices-gp-service",
-    "ogc-api-features": "ogc-features",
-    "ogc_features": "ogc-features",
-    "odata-v4": "odata",
-    "wfs-2.0": "wfs",
-    "wms-1.3.0": "wms",
-    "wmts-1.0.0": "wmts",
-}
-
-CAPABILITIES = (
-    "query",
-    "queryAggregate",
-    "queryExtent",
-    "queryObjectIds",
-    "queryRelated",
-    "applyEdits",
-    "attachments",
-    "render",
-    "tiles",
-    "sql",
-    "stream",
-    "pbf",
-    "connect",
-    "image",
-    "geometry",
-    "geoprocess",
-    "processes",
-)
-
-DEFAULT_CAPABILITIES: Mapping[str, tuple[str, ...]] = {
-    "grpc": (
-        "query",
-        "queryAggregate",
-        "queryExtent",
-        "queryObjectIds",
-        "applyEdits",
-        "stream",
-    ),
-    "geoservices-feature-service": (
-        "query",
-        "queryAggregate",
-        "queryExtent",
-        "queryObjectIds",
-        "queryRelated",
-        "applyEdits",
-        "attachments",
-        "sql",
-        "stream",
-        "pbf",
-        "connect",
-    ),
-    "geoservices-map-service": (
-        "query",
-        "queryAggregate",
-        "queryExtent",
-        "queryObjectIds",
-        "queryRelated",
-        "render",
-        "tiles",
-        "sql",
-        "stream",
-    ),
-    "geoservices-image-service": (
-        "query",
-        "queryExtent",
-        "queryObjectIds",
-        "image",
-        "render",
-        "tiles",
-        "connect",
-    ),
-    "geoservices-geometry-service": ("geometry", "connect"),
-    "geoservices-gp-service": ("geoprocess", "connect"),
-    "ogc-features": ("query", "queryObjectIds", "applyEdits", "stream"),
-    "ogc-tiles": ("render", "tiles"),
-    "ogc-maps": ("render",),
-    "stac": ("query", "queryObjectIds", "stream"),
-    "wfs": ("query", "queryExtent", "queryObjectIds", "applyEdits", "stream"),
-    "wms": ("render", "tiles", "query"),
-    "wmts": ("render", "tiles"),
-    "odata": ("query", "queryObjectIds", "stream", "applyEdits"),
-    "maplibre-vector": ("render", "tiles"),
-    "maplibre-raster": ("render", "tiles"),
-    "maplibre-geojson": ("render",),
-}
-
-_NORMALIZED_PROTOCOL_ALIASES = {key.lower().replace("_", "-"): value for key, value in PROTOCOL_ALIASES.items()}
-_CAPABILITY_BY_KEY = {"".join(ch for ch in capability.lower() if ch.isalnum()): capability for capability in CAPABILITIES}
-
-
-def normalize_protocol(value: Protocol | str) -> str:
-    """Return the canonical cross-SDK protocol id for a protocol name or alias.
-
-    Accepts any string at runtime — the :data:`Protocol` ``Literal`` is the
-    strict static type, and this helper is the explicit validate/coerce
-    boundary for arbitrary input (alias resolution, snake_case, case
-    folding). Raises :class:`ValueError` for unknown protocols.
-    """
-    normalized = str(value).strip().lower().replace("_", "-")
-    protocol = _NORMALIZED_PROTOCOL_ALIASES.get(normalized, normalized)
-    if protocol in PROTOCOLS:
-        return protocol
-    expected = ", ".join(sorted(PROTOCOLS))
-    raise ValueError(f"Unsupported protocol {value!r}. Expected one of: {expected}.")
-
-
-def normalize_capability(value: Capability | str) -> str:
-    """Return the canonical cross-SDK capability id for a capability name.
-
-    Accepts any string at runtime — :data:`Capability` is a strict
-    ``Literal``; this helper is the explicit validate/coerce boundary
-    that resolves snake_case and case variants. Raises :class:`ValueError`
-    for unknown capabilities.
-    """
-    key = "".join(ch for ch in str(value).strip().lower() if ch.isalnum())
-    try:
-        return _CAPABILITY_BY_KEY[key]
-    except KeyError as exc:
-        expected = ", ".join(sorted(CAPABILITIES))
-        raise ValueError(f"Unsupported capability {value!r}. Expected one of: {expected}.") from exc
-
-
-def capability_set(values: Iterable[Capability | str] | None) -> frozenset[str]:
-    """Normalize a capability iterable into canonical capability ids."""
-    if values is None:
-        return frozenset()
-    return frozenset(normalize_capability(value) for value in values)
-
-
-def default_capabilities(protocol: Protocol | str) -> frozenset[str]:
-    """Return the default capabilities advertised for a canonical protocol id."""
-    return frozenset(DEFAULT_CAPABILITIES.get(normalize_protocol(protocol), ()))
-
-
-@dataclass(frozen=True)
-class ServiceSummary:
-    """GeoServices catalog service summary."""
-
-    name: str
-    type: str | None = None
-    url: str | None = None
-    raw: Mapping[str, Any] = field(default_factory=dict)
-
-    @classmethod
-    def from_dict(cls, payload: Mapping[str, Any]) -> "ServiceSummary":
-        return cls(
-            name=str(payload.get("name") or payload.get("serviceName") or ""),
-            type=_optional_str(payload.get("type")),
-            url=_optional_str(payload.get("url")),
-            raw=dict(payload),
-        )
-
-
-@dataclass(frozen=True)
-class DataPlaneCapabilities:
-    """Data-plane capability discovery result."""
-
-    server_version: str | None = None
-    release_channel: str | None = None
-    protocols: frozenset[str] = frozenset()
-    features: Mapping[str, bool] = field(default_factory=dict)
-    raw: Mapping[str, Any] = field(default_factory=dict)
-
-    @classmethod
-    def from_dict(cls, payload: Mapping[str, Any]) -> "DataPlaneCapabilities":
-        features = _capability_flags(_first_present(payload, "features", "featureFlags"))
-        compatibility = payload.get("compatibility")
-        if isinstance(compatibility, Mapping):
-            features = {**_capability_flags(compatibility.get("features")), **features}
-
-        return cls(
-            server_version=_optional_str(_first_present(payload, "serverVersion", "server_version", "version")),
-            release_channel=_optional_str(_first_present(payload, "releaseChannel", "release_channel", "channel")),
-            protocols=frozenset(_capability_names(_first_present(payload, "protocols", "dataProtocols", "surfaces"))),
-            features=features,
-            raw=dict(payload),
-        )
-
-    @classmethod
-    def from_discovery(
-        cls,
-        *,
-        readiness: Mapping[str, Any],
-        catalog: Mapping[str, Any],
-    ) -> "DataPlaneCapabilities":
-        protocols: set[str] = set(_capability_names(readiness.get("protocols")))
-        services = catalog.get("services")
-        if isinstance(services, Sequence) and not isinstance(services, str):
-            protocols.add("geoservices")
-            for service in services:
-                if not isinstance(service, Mapping):
-                    continue
-                service_type = _optional_str(service.get("type"))
-                if service_type is not None:
-                    protocols.add(_normalize_advertised_name(service_type))
-
-        return cls(
-            server_version=_optional_str(_first_present(readiness, "serverVersion", "server_version", "version")),
-            release_channel=_optional_str(_first_present(readiness, "releaseChannel", "release_channel", "channel")),
-            protocols=frozenset(protocols),
-            features={
-                "readiness": bool(readiness),
-                "service-catalog": isinstance(services, Sequence) and not isinstance(services, str),
-            },
-            raw={"readiness": dict(readiness), "catalog": dict(catalog)},
-        )
-
-    def supports(self, capability: str) -> bool:
-        """Return whether a named protocol or feature is advertised."""
-        keys = _normalized_surface_keys(capability)
-        return any(key in self.protocols or bool(self.features.get(key, False)) for key in keys)
-
-
-@dataclass(frozen=True)
-class SourceLocator:
-    """Protocol-specific source address using Pythonic field names.
-
-    Attributes:
-        service_id: GeoServices service identifier (FeatureServer/MapServer/ImageServer).
-        layer_id: Numeric layer index within a GeoServices service.
-        collection_id: OGC API / STAC collection identifier.
-        entity_set: OData entity-set name (e.g. ``"Features"``).
-        type_name: WFS ``typeName`` value.
-    """
-
-    service_id: str | None = None
-    layer_id: int | None = None
-    collection_id: str | None = None
-    entity_set: str | None = None
-    type_name: str | None = None
-
-    @classmethod
-    def from_dict(cls, payload: Mapping[str, Any]) -> "SourceLocator":
-        return cls(
-            service_id=_optional_str(_first_present(payload, "serviceId", "service_id")),
-            layer_id=_optional_int(_first_present(payload, "layerId", "layer_id")),
-            collection_id=_optional_str(_first_present(payload, "collectionId", "collection_id")),
-            entity_set=_optional_str(_first_present(payload, "entitySet", "entity_set")),
-            type_name=_optional_str(_first_present(payload, "typeName", "type_name")),
-        )
-
-
-@dataclass(frozen=True)
-class SourceDescriptor:
-    """Cross-SDK source description used by the source facade.
-
-    ``protocol`` accepts any string (including alias forms) at construction
-    time; ``__post_init__`` normalizes it through :func:`normalize_protocol`
-    so the stored value is always a canonical :data:`Protocol` literal.
-
-    Attributes:
-        id: Stable source identifier used by the canonical facade.
-        protocol: Canonical protocol literal after normalization.
-        locator: Protocol-specific addressing fields.
-        capabilities: Frozen set of canonical capability names advertised.
-        raw: Free-form mapping preserving the source's underlying payload.
-    """
-
-    id: str
-    protocol: Protocol | str
-    locator: SourceLocator = field(default_factory=SourceLocator)
-    capabilities: frozenset[str] = field(default_factory=frozenset)
-    raw: Mapping[str, Any] = field(default_factory=dict)
-
-    def __post_init__(self) -> None:
-        protocol = normalize_protocol(self.protocol)
-        capabilities = capability_set(self.capabilities) or default_capabilities(protocol)
-        object.__setattr__(self, "protocol", protocol)
-        object.__setattr__(self, "capabilities", capabilities)
-        object.__setattr__(self, "raw", dict(self.raw))
-
-    @classmethod
-    def from_dict(cls, payload: Mapping[str, Any]) -> "SourceDescriptor":
-        locator = payload.get("locator")
-        return cls(
-            id=str(payload.get("id") or ""),
-            protocol=_optional_str(payload.get("protocol")) or "geoservices-feature-service",
-            locator=SourceLocator.from_dict(locator) if isinstance(locator, Mapping) else SourceLocator(),
-            capabilities=capability_set(_sequence_value(payload.get("capabilities"))),
-            raw=dict(payload),
-        )
-
-    def supports(self, capability: Capability | str) -> bool:
-        """Return whether the source descriptor advertises a capability."""
-        return normalize_capability(capability) in self.capabilities
-
-
-@dataclass(frozen=True)
-class Pagination:
-    """Pagination options shared by queryable source protocols.
-
-    Attributes:
-        limit: Maximum features to return across all pages.
-        page_size: Per-request page size for paginated protocols.
-        max_pages: Cap on the number of pages walked.
-        offset: Starting offset for the first page (when supported).
-    """
-
-    limit: int | None = None
-    page_size: int | None = None
-    max_pages: int | None = None
-    offset: int | None = None
-
-    @classmethod
-    def from_dict(cls, payload: Mapping[str, Any]) -> "Pagination":
-        return cls(
-            limit=_optional_int(payload.get("limit")),
-            page_size=_optional_int(_first_present(payload, "pageSize", "page_size")),
-            max_pages=_optional_int(_first_present(payload, "maxPages", "max_pages")),
-            offset=_optional_int(payload.get("offset")),
-        )
-
-
-@dataclass(frozen=True)
-class Query:
-    """Cross-SDK query model with Pythonic field names.
-
-    Filter routing
-    --------------
-
-    Pick exactly one of three forms, matched to the target protocol:
-
-    * ``where`` — SQL-style ``WHERE`` clause for SQL protocols
-      (GeoServices FeatureServer, OData). On CQL-based protocols (OGC
-      Features, STAC) this raises :class:`ValueError` at routing time,
-      because silently forwarding SQL syntax to a CQL endpoint is a
-      footgun that masks bugs.
-    * ``cql_filter`` — CQL2-text filter for CQL protocols (OGC Features,
-      STAC). On SQL-style protocols this raises :class:`ValueError` —
-      CQL2-text is not valid for FeatureServer / OData.
-    * ``where_as_cql=True`` — escape hatch for protocol-agnostic callers
-      that have already verified the ``where`` string is valid CQL2-text.
-      With the flag set, ``where`` is forwarded to the CQL ``filter``
-      field on OGC/STAC without raising. The flag is a no-op on SQL-style
-      protocols (``where`` still routes to SQL ``where``).
-
-    When both ``where`` and ``cql_filter`` are set on a CQL-based
-    protocol, ``cql_filter`` wins.
-
-    Attributes:
-        where: SQL-style filter for FeatureServer/OData endpoints.
-        cql_filter: CQL2-text filter for OGC Features / STAC.
-        where_as_cql: When True, forwards ``where`` to CQL endpoints without raising.
-        spatial_filter: Free-form spatial-filter mapping (geometry, relation, SR).
-        bbox: ``(minx, miny, maxx, maxy)`` spatial filter; comma-string also accepted.
-        out_fields: Field selector; ``["*"]`` or ``"*"`` selects all.
-        order_by: Sort specification forwarded to the protocol.
-        pagination: :class:`Pagination` options (``limit``, ``page_size``, ...).
-        aggregation: Protocol-neutral aggregation request mapping.
-        return_geometry: Whether to include geometry in the response.
-        out_sr: Output spatial reference (EPSG code or WKID).
-        extra_params: Free-form per-protocol query parameter overrides.
-    """
-
-    where: str | None = None
-    cql_filter: str | None = None
-    where_as_cql: bool = False
-    spatial_filter: Mapping[str, Any] | None = None
-    bbox: str | Sequence[int | float] | None = None
-    out_fields: str | Sequence[str] | None = None
-    order_by: str | Sequence[str] | None = None
-    pagination: Pagination = field(default_factory=Pagination)
-    aggregation: Mapping[str, Any] | None = None
-    return_geometry: bool = True
-    out_sr: int | str | None = None
-    extra_params: Mapping[str, Any] = field(default_factory=dict)
-
-    def __post_init__(self) -> None:
-        pagination = self.pagination
-        if isinstance(pagination, Mapping):
-            pagination = Pagination.from_dict(pagination)
-        object.__setattr__(self, "pagination", pagination)
-        object.__setattr__(self, "spatial_filter", dict(self.spatial_filter) if self.spatial_filter else None)
-        object.__setattr__(self, "aggregation", dict(self.aggregation) if self.aggregation else None)
-        object.__setattr__(self, "extra_params", dict(self.extra_params))
-
-    @classmethod
-    def from_dict(cls, payload: Mapping[str, Any]) -> "Query":
-        pagination = _first_present(payload, "pagination", "page")
-        return cls(
-            where=_optional_str(payload.get("where")),
-            cql_filter=_optional_str(_first_present(payload, "cqlFilter", "cql_filter")),
-            where_as_cql=bool(_first_present(payload, "whereAsCql", "where_as_cql") or False),
-            spatial_filter=_mapping_value(_first_present(payload, "spatialFilter", "spatial_filter")),
-            bbox=_first_present(payload, "bbox", "boundingBox", "bounding_box"),
-            out_fields=_first_present(payload, "outFields", "out_fields", "fields"),
-            order_by=_first_present(payload, "orderBy", "order_by"),
-            pagination=Pagination.from_dict(pagination) if isinstance(pagination, Mapping) else Pagination(),
-            aggregation=_mapping_value(payload.get("aggregation")),
-            return_geometry=bool(_first_present(payload, "returnGeometry", "return_geometry"))
-            if _first_present(payload, "returnGeometry", "return_geometry") is not None
-            else True,
-            out_sr=_first_present(payload, "outSr", "out_sr"),
-            extra_params=_mapping_value(_first_present(payload, "extraParams", "extra_params")) or {},
-        )
-
-    @property
-    def limit(self) -> int | None:
-        return self.pagination.limit
-
-    @property
-    def page_size(self) -> int | None:
-        return self.pagination.page_size
-
-    @property
-    def max_pages(self) -> int | None:
-        return self.pagination.max_pages
-
-    @property
-    def offset(self) -> int | None:
-        return self.pagination.offset
-
-
-@dataclass(frozen=True)
-class DegradedReason:
-    """Reason a result used a lower-fidelity path for a requested capability.
-
-    ``capability`` and ``protocol`` are normalized to their canonical
-    Literal form in ``__post_init__`` via :func:`normalize_capability` /
-    :func:`normalize_protocol`, so aliases passed at construction time are
-    accepted but stored values always match the strict types.
-    """
-
-    capability: Capability | str
-    protocol: Protocol | str
-    source_id: str | None = None
-    reason: str = ""
-
-    def __post_init__(self) -> None:
-        object.__setattr__(self, "capability", normalize_capability(self.capability))
-        object.__setattr__(self, "protocol", normalize_protocol(self.protocol))
-
-
-@dataclass(frozen=True)
-class Result(Generic[T]):
-    """Collected result returned by the canonical Source/Query API.
-
-    Generic over the feature element type ``T`` (defaults to
-    :class:`QueryFeature` at call sites). ``raw_legacy`` exposes the
-    underlying :class:`FeatureQueryResult` (the protocol-neutral query
-    result the canonical facade is built on) for callers that need the
-    unprocessed protocol response — e.g. inspecting ``pages_seen`` or the
-    original ``query`` envelope. ``raw`` remains a free-form mapping
-    reserved for protocol-specific extension fields.
-
-    Attributes:
-        features: Tuple of result features (type parameter ``T``).
-        exceeded_transfer_limit: True when the server signalled more pages remain.
-        total_count: Server-reported total count when available.
-        aggregate_rows: Aggregate result rows when an aggregation was requested.
-        extent: Result extent mapping (``xmin``/``ymin``/``xmax``/``ymax``).
-        fields: Field schema entries returned by the protocol.
-        degraded: Reasons this result fell back to a lower-fidelity path.
-        protocol: Canonical protocol literal the result was served from.
-        source_id: Identifier of the source that produced the result.
-        query: The :class:`Query` instance that produced this result.
-        raw: Free-form mapping for protocol-specific extension fields.
-        raw_legacy: Underlying :class:`FeatureQueryResult` envelope, when available.
-    """
-
-    features: tuple[T, ...] = ()
-    exceeded_transfer_limit: bool = False
-    total_count: int | None = None
-    aggregate_rows: tuple[Mapping[str, Any], ...] = ()
-    extent: Mapping[str, Any] | None = None
-    fields: tuple[Mapping[str, Any], ...] = ()
-    degraded: tuple[DegradedReason, ...] = ()
-    protocol: str = ""
-    source_id: str = ""
-    query: Query | None = None
-    raw: Mapping[str, Any] = field(default_factory=dict)
-    raw_legacy: "FeatureQueryResult | None" = None
-
-    def __post_init__(self) -> None:
-        object.__setattr__(self, "features", tuple(self.features))
-        object.__setattr__(self, "aggregate_rows", tuple(dict(row) for row in self.aggregate_rows))
-        object.__setattr__(self, "fields", tuple(dict(field) for field in self.fields))
-        object.__setattr__(self, "extent", dict(self.extent) if self.extent is not None else None)
-        object.__setattr__(self, "degraded", tuple(self.degraded))
-        object.__setattr__(self, "raw", dict(self.raw))
-
-
-@dataclass(frozen=True)
-class Feature:
-    """FeatureServer feature with attributes and optional geometry.
-
-    Returned by GeoServices FeatureServer endpoints (``FeatureServerClient``
-    and its async sibling) via :meth:`FeatureSet.features`. Uses the
-    GeoServices ``attributes``/``geometry`` shape verbatim. For the
-    protocol-neutral feature shape used by :meth:`Source.query`, see
-    :class:`QueryFeature` instead.
-    """
-
-    attributes: Mapping[str, Any]
-    geometry: Mapping[str, Any] | None = None
-    raw: Mapping[str, Any] = field(default_factory=dict)
-
-    @classmethod
-    def from_dict(cls, payload: Mapping[str, Any]) -> "Feature":
-        attributes = payload.get("attributes")
-        geometry = payload.get("geometry")
-        return cls(
-            attributes=dict(attributes) if isinstance(attributes, Mapping) else {},
-            geometry=dict(geometry) if isinstance(geometry, Mapping) else None,
-            raw=dict(payload),
-        )
-
-    @property
-    def object_id(self) -> int | None:
-        for key in ("objectid", "objectId", "OBJECTID"):
-            value = self.attributes.get(key)
-            if value is not None:
-                return int(value)
-        return None
-
-
-@dataclass(frozen=True)
-class FeatureSet:
-    """Typed FeatureServer query response."""
-
-    features: tuple[Feature, ...]
-    fields: tuple[Mapping[str, Any], ...] = ()
-    geometry_type: str | None = None
-    spatial_reference: Mapping[str, Any] | None = None
-    exceeded_transfer_limit: bool = False
-    raw: Mapping[str, Any] = field(default_factory=dict)
-
-    @classmethod
-    def from_dict(cls, payload: Mapping[str, Any]) -> "FeatureSet":
-        raw_features = payload.get("features")
-        raw_fields = payload.get("fields")
-        spatial_reference = payload.get("spatialReference")
-        return cls(
-            features=tuple(
-                Feature.from_dict(feature)
-                for feature in raw_features or []
-                if isinstance(feature, Mapping)
-            ),
-            fields=tuple(dict(field) for field in raw_fields or [] if isinstance(field, Mapping)),
-            geometry_type=_optional_str(payload.get("geometryType")),
-            spatial_reference=dict(spatial_reference) if isinstance(spatial_reference, Mapping) else None,
-            exceeded_transfer_limit=bool(payload.get("exceededTransferLimit", False)),
-            raw=dict(payload),
-        )
-
-
-@dataclass(frozen=True)
-class FeatureQuery:
-    """Protocol-neutral feature query request."""
-
-    source: str
-    protocol: QueryProtocol | str = "feature-server"
-    layer_id: int | None = None
-    where: str | None = None
-    filter: str | None = None
-    bbox: str | Sequence[int | float] | None = None
-    fields: str | Sequence[str] | None = None
-    return_geometry: bool = True
-    page_size: int | None = None
-    limit: int | None = None
-    max_pages: int | None = None
-    extra_params: Mapping[str, Any] = field(default_factory=dict)
-
-
-@dataclass(frozen=True)
-class QueryFeature:
-    """Protocol-neutral feature returned by the shared query API.
-
-    Returned by :meth:`Source.query`, :meth:`Source.stream`, and the
-    underlying :meth:`HonuaClient.query` / :meth:`AsyncHonuaClient.query`
-    facade across every protocol (FeatureServer, OGC Features, STAC,
-    OData). Uses GeoJSON-shaped ``properties`` + ``geometry``. For the
-    raw GeoServices ``attributes``/``geometry`` shape, see :class:`Feature`.
-
-    Attributes:
-        id: Stable feature identifier from the underlying protocol.
-        properties: GeoJSON-shaped attribute mapping.
-        geometry: Optional GeoJSON-shaped geometry mapping.
-        protocol: Canonical protocol literal the feature was served from.
-        source: Identifier of the source that produced the feature.
-        raw: Free-form mapping preserving the underlying protocol payload.
-    """
-
-    id: str | int | None
-    properties: Mapping[str, Any]
-    geometry: Mapping[str, Any] | None = None
-    protocol: str = ""
-    source: str = ""
-    raw: Mapping[str, Any] = field(default_factory=dict)
-
-
-@dataclass(frozen=True)
-class FeatureQueryResult:
-    """Collected result returned by the shared query API.
-
-    Pagination signals (``exceeded_transfer_limit``, ``total_count``,
-    ``pages_seen``) are populated from the underlying protocol response
-    when available:
-
-    * GeoServices FeatureServer surfaces ``exceededTransferLimit`` on each
-      page; ``total_count`` defaults to ``len(features)``.
-    * OGC Features / STAC surface ``numberMatched`` (total) and
-      ``numberReturned``; ``exceeded_transfer_limit`` is derived from a
-      ``next`` link being present on the last page walked.
-    * OData surfaces ``@odata.count`` (total) and ``@odata.nextLink``
-      (drives ``exceeded_transfer_limit``).
-
-    When the protocol does not expose a signal the field defaults to a
-    safe value (``False`` / ``None``) — never silently fabricated.
-    """
-
-    features: tuple[QueryFeature, ...]
-    protocol: str
-    source: str
-    query: FeatureQuery
-    exceeded_transfer_limit: bool = False
-    total_count: int | None = None
-    pages_seen: int = 0
-
-
-@dataclass(frozen=True)
-class EditOperationResult:
-    """One add, update, or delete result from applyEdits."""
-
-    success: bool
-    object_id: int | None = None
-    global_id: str | None = None
-    error: Mapping[str, Any] | None = None
-    raw: Mapping[str, Any] = field(default_factory=dict)
-
-    @classmethod
-    def from_dict(cls, payload: Mapping[str, Any]) -> "EditOperationResult":
-        return cls(
-            success=bool(payload.get("success", False)),
-            object_id=_optional_int(_first_present(payload, "objectId", "objectid")),
-            global_id=_optional_str(_first_present(payload, "globalId", "globalid")),
-            error=dict(payload["error"]) if isinstance(payload.get("error"), Mapping) else None,
-            raw=dict(payload),
-        )
-
-
-@dataclass(frozen=True)
-class ApplyEditsResult:
-    """Typed applyEdits response grouped by operation."""
-
-    add_results: tuple[EditOperationResult, ...] = ()
-    update_results: tuple[EditOperationResult, ...] = ()
-    delete_results: tuple[EditOperationResult, ...] = ()
-    raw: Mapping[str, Any] = field(default_factory=dict)
-
-    @classmethod
-    def from_dict(cls, payload: Mapping[str, Any]) -> "ApplyEditsResult":
-        return cls(
-            add_results=_edit_results(payload.get("addResults")),
-            update_results=_edit_results(payload.get("updateResults")),
-            delete_results=_edit_results(payload.get("deleteResults")),
-            raw=dict(payload),
-        )
-
-    @property
-    def all_succeeded(self) -> bool:
-        results = [*self.add_results, *self.update_results, *self.delete_results]
-        return bool(results) and all(result.success for result in results)
-
-
-def _edit_results(value: Any) -> tuple[EditOperationResult, ...]:
-    if not isinstance(value, Sequence) or isinstance(value, str):
-        return ()
-    return tuple(EditOperationResult.from_dict(item) for item in value if isinstance(item, Mapping))
-
-
-def _sequence_value(value: Any) -> Sequence[Any] | None:
-    if isinstance(value, Sequence) and not isinstance(value, str):
-        return value
-    return None
-
-
-def _mapping_value(value: Any) -> Mapping[str, Any] | None:
-    if isinstance(value, Mapping):
-        return dict(value)
-    return None
-
-
-def _optional_str(value: Any) -> str | None:
-    return str(value) if value is not None else None
-
-
-def _optional_int(value: Any) -> int | None:
-    return int(value) if value is not None else None
-
-
-def _first_present(payload: Mapping[str, Any], *keys: str) -> Any:
-    for key in keys:
-        if key in payload:
-            return payload[key]
-    return None
-
-
-def _capability_flags(value: Any) -> dict[str, bool]:
-    if not isinstance(value, Mapping):
-        return {}
-    return {_normalize_advertised_name(str(key)): bool(flag) for key, flag in value.items()}
-
-
-def _capability_names(value: Any) -> set[str]:
-    if not isinstance(value, Sequence) or isinstance(value, str):
-        if isinstance(value, Mapping):
-            return {
-                _normalize_advertised_name(str(key))
-                for key, enabled in value.items()
-                if _capability_enabled(enabled)
-            }
-        return set()
-
-    names: set[str] = set()
-    for item in value:
-        if isinstance(item, str):
-            names.add(_normalize_advertised_name(item))
-            continue
-        if not isinstance(item, Mapping) or not _capability_enabled(item):
-            continue
-        name = _first_present(item, "id", "name", "protocol", "surface")
-        if name is not None:
-            names.add(_normalize_advertised_name(str(name)))
-    return names
-
-
-def _capability_enabled(value: Any) -> bool:
-    if isinstance(value, Mapping):
-        enabled = value.get("enabled", True)
-        return bool(enabled)
-    return bool(value)
-
-
-def _normalize_capability_name(value: str) -> str:
-    normalized = value.strip().lower().replace("_", "-").replace("/", "-").replace(" ", "-")
-    aliases = {
-        "feature-server": "featureserver",
-        "feature-service": "featureserver",
-        "geo-services": "geoservices",
-        "geocode-server": "geocodeserver",
-        "geocoding": "geocodeserver",
-        "geometry-server": "geometryserver",
-        "image-server": "imageserver",
-        "map-server": "mapserver",
-        "ogc-features": "ogc-features",
-        "ogc-api-features": "ogc-features",
-        "ogc-maps": "ogc-maps",
-        "ogc-api-maps": "ogc-maps",
-        "ogc-tiles": "ogc-tiles",
-        "ogc-api-tiles": "ogc-tiles",
-        "ogc-coverages": "ogc-coverages",
-        "ogc-api-coverages": "ogc-coverages",
-        "ogc-processes": "ogc-processes",
-        "ogc-api-processes": "ogc-processes",
-    }
-    return aliases.get(normalized, normalized)
-
-
-def _normalize_advertised_name(value: str) -> str:
-    try:
-        return normalize_protocol(value)
-    except ValueError:
-        pass
-    try:
-        return normalize_capability(value)
-    except ValueError:
-        return _normalize_capability_name(value)
-
-
-def _normalized_surface_keys(value: str) -> set[str]:
-    keys = {_normalize_capability_name(value)}
-    with contextlib.suppress(ValueError):
-        keys.add(normalize_protocol(value))
-    with contextlib.suppress(ValueError):
-        keys.add(normalize_capability(value))
-    return keys
